Separated types of class and method into different files, prevented debrevented unnecessary debug-printing of key-value pair dicts
